DISPLAY THE ARRAY OBJECTS ONTO WEBPGAE
I have added a div container onto the HTML file and this container will hold all the cards displaying the books. Further, I used the forEach() method to loop through the array, created a new div with a class of card, and create a multi-line string called content that containes the HTML Markup for each card layout that will display each book on its own card on the webpage.