ADD TOGGLE BUTTON FOR READ STATUS OF EACH BOOK
I have created a toggle button on the bookDisplay function and added an event listener to the same. Further, I have created a function to toggle the buttons form read to unread status for each book. This required updating the Book constructor to include a read property with a default of false. The button now works as expected.