ADD FORM VALIDATION VIA JAVASCRIPT
I have changed the previous code to add form validation in javascript. I have added a showError function that shows a custom validation message to the user when they fail to input anything on any of the input fields. Further, I have added CSS styling that changes color of input fields when there is no input. The form will not submit if any input field is invalid.